WIP: Initial table comparing MS Leader topologies for Cody
This table fell out of work on the "About-multi-site" page; we
concluded that (1) docs need a table comparing the two topologies that
Cody might select as Leader, enumerating their specific (dis)advantages
(e.g. "HA leader stays online during stemcell upgrades, SN-MS leader
does not), and (2) Cody is the audience for that table, therefore it
belongs on this "Using multi-site" page.

The table as-is is really just a placeholder, copied from
"About-multi-site" work.

[#186529355](https://www.pivotaltracker.com/story/show/186529355)

Co-authored-by: Kim Bassett <kbassett@vmware.com>
Co-authored-by: Kyle Ong <kyleo@vmware.com>